Aus region selector - Fix linting errors
